worker/requirements.txt
# LangGraph Worker dependencies
langgraph
langchain-anthropic
anthropic
asyncpg
upstash-redis
boto3
pymupdf
python-pptx
